neuralDomains/requirements.txt
python_environ
Django
environ
gunicorn
django-crispy-forms
psycopg2-binary
whitenoise
python-dotenv